描画処理関数切り出し
